# attached_assets-4/attached_assets/2/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Clientum CRM/ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List quotes
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Not found
    patch:
      operationId: updateQuote
      tags: [quotes]
      summary: Update quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Not found
    delete:
      operationId: deleteQuote
      tags: [quotes]
      summary: Delete quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    QuoteItem:
      type: object
      required: [name, quantity, unitPrice, total]
      properties:
        productId:
          type: integer
        name:
          type: string
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
    Quote:
      type: object
      required: [id, number, title, status, items, subtotal, discount, total, createdAt]
      properties:
        id:
          type: integer
        number:
          type: string
        contactId:
          type: ["integer", "null"]
        contactName:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        title:
          type: string
        status:
          type: string
          enum: [draft, sent, accepted, rejected, expired]
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        validUntil:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    QuoteInput:
      type: object
      required: [title]
      properties:
        contactId:
          type: integer
        companyId:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [draft, sent, accepted, rejected, expired]
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        discount:
          type: number
        validUntil:
          type: string
        notes:
          type: string
    QuoteUpdate:
      type: object
      properties:
        contactId:
          type: integer
        companyId:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [draft, sent, accepted, rejected, expired]
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        discount:
          type: number
        validUntil:
          type: string
        notes:
          type: string
...